+++ src/app/np-input-text-demo/np-input-text-demo-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { NpInputTextDemoComponent } from './np-input-text-demo.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: NpInputTextDemoComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class NpInputTextDemoRoutingModule { }