+++ src/app/np-utility-demo/np-utility-demo.module.ts
@@ -5,7 +5,7 @@ import { NpUtilityDemoRoutingModule } from './np-utility-demo-routing.module';
 import { NpUtilityDemoComponent } from './np-utility-demo.component';
 import {
   NpUtilityModule, NpHightLightPipe, NpDatePickerModule, NpTimePickerModule, NpSwitchModule,
-  NpFileUploadModule, NpCardModule
+  NpFileUploadModule, NpCardModule, NpInputTextModule
 } from 'np-ui-lib';
 import { FormsModule } from '@angular/forms';
 
@@ -18,6 +18,7 @@ import { FormsModule } from '@angular/forms';
     NpUtilityModule,
     NpDatePickerModule,
     NpTimePickerModule,
+    NpInputTextModule,
     NpSwitchModule,
     NpFileUploadModule,
     NpCardModule