+++ projects/np-ui-lib/CHANGELOG.md
@@ -19,6 +19,7 @@
 - Change onSelect and onDeselect parameters in datagrid
 - Add new radio button component
 - Remove clear button from datepicker, timepicker, colorpicker, autocomplete, tags
+- Add new input text component
 
 ## 10.5.2
 - Add getCurrentStateName Api to data grid